+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * Adds layer.Conversation.leave() to leave a Conversation and remove it from my Conversation list.  New Messages will NOT cause the Conversation to reappear.
 * layer.Message
   * Support for Deletion with either layer.Constants.DELETION_MODE.MY_DEVICES or .Constants.DELETION_MODE.ALL (delete for all users or just for me).  Note that deleting for just me doesn't remove me as a participant, which means that new Messages will cause the Conversation to reappear.
+  * The `messages:read` event has been removed. Please use the `messages:change` event instead.
 
 ## 1.0.0